+++ CHANGELOG.markdown
@@ -2,7 +2,7 @@
 =====
 
 * Fixed serious bug for default column names.  record.middle_name instead of record.middlename
-
+* Spreadsheet can be re-read from file by calling @sheet.reload_spreadsheet(opts)
 
 
 0.0.3